Add results oriented interface
In addition to toBuffer and fromBuffer, we now support toBufferResult
and fromBufferResult, which return Result {error, value} objects.

Borrows the Result object introduced in BufRW.
This Result constructor just takes error and value arguments, reflecting
the callback argument order, and eliminates the use of just and error
constructor functions for a modest performance gain.

Retains the error context feature by annotating the error side of a
result on the way out of uglify and reify stack frames.